requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
